# _uvoz/bookie-uvoz-prejemnic/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ERP — Modul prevzema blaga
  description: |
    Uvoz prejemnic in dobavnic dobaviteljev za POS gostinstvo.
    Vsi kanali uvoza (datoteka, e-pošta, Peppol, API, OCR) se stekajo
    v isti tok obdelave: zajem → razčlenitev → uparjanje → validacija → knjiženje.
  version: "1.0.0"

# ⚠ OPOMBA: ta datoteka je bila pisana kot pogodba pred izvedbo.
# Dejanske poti so /api/pos/uvoz/... in so definirane v
# artifacts/api-server/src/routes/pos/uvoz-prejemnic.ts in uvoz-profili.ts.
# Ob razhajanju velja koda. Datoteka ostaja kot pregled obsega vmesnika.
servers:
  - url: /api/pos/uvoz
security:
  - bearerAuth: []
...
paths:
...
  /preslikave/{preslikavaId}:
    delete:
      tags: [Uparjanje]
      summary: Odstrani preslikavo
      parameters:
        - name: preslikavaId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Odstranjeno }

  # ==================== USKLAJEVANJE ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...